Correct hatcheries displayed and downloads accordingly, graph now shows data
